Updated QLBookingBLL + RoomtypeDAL Form_Detail_Invoice: Updated UI & Function Export to PDF Form_Detail_Booking: Updated Enabling of btn_Checkin

+++ PBL3REAL/View/Form_Detail_Room_Category.cs
@@ -234,7 +234,10 @@ namespace PBL3REAL.View
             }
             try
             {
-                File.Delete(fullpath);
+                if (File.Exists(fullpath))
+                {
+                    File.Delete(fullpath);
+                }   
             }
             catch (Exception e) { }
         }
